# Improve top-k sampling CUDA kernel