Throughput numbers (tok/sec) for Qwen-2.5-1.5B int4 model on RTX 4090 with different top-K values and fixed top-p=0.8 value.

K | Before | After | Speed-up
-- | -- | -- | --
1 | 210 | 217 | 1.03
20 | 178 | 206 | 1.16
50 | 141 | 195 | 1.38

Top-K kernel shows an improvement of ~5.7x for top-k = 50 and 13x for top-k = 1.

Nsight profile before the change for Top-k = 50:

<img width="1154" height="696" alt="image" src="https://github.com/user-attachments/assets/ba50bae8-bb1a-4658-a422-67d0cf5f186d" />


Nsight profile after the change for Top-K = 50:

<img width="935" height="739" alt="image" src="https://github.com/user-attachments/assets/6631015f-a191-4d8b-9087-3b4d17070fa0" />


Nsight profile before the change for Top-k = 1:
<img width="1102" height="743" alt="image" src="https://github.com/user-attachments/assets/57776d74-53c8-43e7-b56d-4e898a1bddd9" />


Nsight profile after the change for Top-k = 1:

<img width="1412" height="760" alt="image" src="https://github.com/user-attachments/assets/b5a83ddc-a74d-44eb-8e9e-24fdf2700605" />