@tianleiwu 2025-09-02
> @tianleiwu the CI/CD failure with MacOS seems like an issue with CI/CD infra. Could you please check?

@gaugarg-nv, 
You need integrate/rebase latest main, which has a commit that fix the build error: https://github.com/microsoft/onnxruntime-genai/commit/e22739fae3c7fbbeba3f485a363f8b80a5ca73c8