@gaugarg-nv 2025-09-03
> > @tianleiwu the CI/CD failure with MacOS seems like an issue with CI/CD infra. Could you please check?
> 
> @gaugarg-nv, You need merge/rebase latest main, which has a commit that fix the build error: [e22739f](https://github.com/microsoft/onnxruntime-genai/commit/e22739fae3c7fbbeba3f485a363f8b80a5ca73c8)

Changes are now rebased on top of main.